Strip `refs/heads` from the branch name passed to shipitscript
Before this, we would be passing the full ref name as the branch name
which would lead to weirdness like this in shipit:

```
  [
    "141.0b1",
    "release/v141"
  ],
  [
    "141.0b2",
    "refs/heads/release/v141"
  ],
```

In that example, b1 got created manually through the UI, b2 got
automatically trigged by the cron. The cronjob then tries to find the
last release on the branch is currently on and decides that there's none
since there's no matching branch. In practice, this isn't much of an
issue, it'll just consider that there's no previous release and promote
the commit as-is without checking if there was any relevant changes
since b1.