Remove `adjust.deeplink_received` event from the metrics ping
This will soon be linted by glean-parser and thus become a warning.
Events on the metrics ping are rarely useful.
None of our automatic tooling assumes there are events on the `metrics` ping.
No other events are on the `metrics` ping, thus no event timeline can be
established anyway.

This event is also the only event on the `first-session` ping, so it's
not clear how useful the event is there either.
However the `first-session` is a custom ping, so it's not getting linted.
It also has a test for it specifically, so I'm keeping it for now.